API: Create new sourcefile
that houses the new public API using the `to_string` and `from_string`
functions and friends. These aren't exported since their names are
pretty nondescript by themselves, so we're reimporting the in the base
module to make them accessible there instead.
